bfa-service/pyproject.toml
[project]
name = "build-failure-analyzer"
version = "0.1.0"
description = "RAG-powered build failure analyzer with Slack integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.2.1",
    "fastapi>=0.121.0",
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-ollama>=1.0.0",
    "pydantic>=2.12.3",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "slack-sdk>=3.37.0",
    "uvicorn[standard]>=0.38.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src", "scripts"]
omit = ["tests/*", "test_*.py", "__pycache__/*", "venv/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]
precision = 2
show_missing = true
